Ignore failed deletions of non-existing watchers.
One-shot watches get deleted on our end when they are dropped, but we
sometimes still get a notification for the watches from zookeeper after
the watchers are dropped.

In those cases, we send a RemoveWatch request to catch any leaking
persistent watchers, but since the watcher does not exist, it results in
a NoWatcher error.

Workaround for #22